# Move ip, ips and hostname to getters in request