--- a/lib/request.js
+++ b/lib/request.js
@@ -1,8 +1,9 @@
 'use strict'
 
+const proxyAddr = require('proxy-addr')
 const { emitWarning } = require('./warnings')
 
-function Request (id, params, req, query, headers, log, ip, ips, hostname) {
+function Request (id, params, req, query, headers, log) {
   this.id = id
   this.params = params
   this.raw = req
@@ -10,13 +11,38 @@ function Request (id, params, req, query, headers, log, ip, ips, hostname) {
   this.headers = headers
   this.log = log
   this.body = null
-  this.ip = ip
-  this.ips = ips
-  this.hostname = hostname
 }
 
-function buildRequest (R) {
-  function _Request (id, params, req, query, headers, log, ip, ips, hostname) {
+function getTrustProxyFn (tp) {
+  if (typeof tp === 'function') {
+    return tp
+  }
+  if (tp === true) {
+    // Support plain true/false
+    return function () { return true }
+  }
+  if (typeof tp === 'number') {
+    // Support trusting hop count
+    return function (a, i) { return i < tp }
+  }
+  if (typeof tp === 'string') {
+    // Support comma-separated tps
+    const vals = tp.split(',').map(it => it.trim())
+    return proxyAddr.compile(vals)
+  }
+  return proxyAddr.compile(tp || [])
+}
+
+function buildRequest (R, trustProxy) {
+  if (trustProxy) {
+    return buildRequestWithTrustProxy(R, trustProxy)
+  }
+
+  return buildRegularRequest(R)
+}
+
+function buildRegularRequest (R) {
+  function _Request (id, params, req, query, headers, log) {
     this.id = id
     this.params = params
     this.raw = req
@@ -24,11 +50,36 @@ function buildRequest (R) {
     this.headers = headers
     this.log = log
     this.body = null
-    this.ip = ip
-    this.ips = ips
-    this.hostname = hostname
   }
   _Request.prototype = new R()
+
+  return _Request
+}
+
+function buildRequestWithTrustProxy (R, trustProxy) {
+  const _Request = buildRegularRequest(R)
+  const proxyFn = getTrustProxyFn(trustProxy)
+
+  Object.defineProperties(_Request.prototype, {
+    ip: {
+      get () {
+        return proxyAddr(this.req, proxyFn)
+      }
+    },
+    ips: {
+      get () {
+        return proxyAddr.all(this.req, proxyFn)
+      }
+    },
+    hostname: {
+      get () {
+        if (this.ip !== undefined && this.headers['x-forwarded-host']) {
+          return this.headers['x-forwarded-host']
+        }
+        return this.headers.host || this.headers[':authority']
+      }
+    }
+  })
 
   return _Request
 }
@@ -53,6 +104,21 @@ Object.defineProperties(Request.prototype, {
   connection: {
     get () {
       return this.raw.connection
+    }
+  },
+  ip: {
+    get () {
+      return this.connection.remoteAddress
+    }
+  },
+  ips: {
+    get () {
+      return undefined
+    }
+  },
+  hostname: {
+    get () {
+      return this.headers.host || this.headers[':authority']
     }
   }
 })