--- a/lib/request.js
+++ b/lib/request.js
@@ -111,11 +111,6 @@ Object.defineProperties(Request.prototype, {
       return this.connection.remoteAddress
     }
   },
-  ips: {
-    get () {
-      return undefined
-    }
-  },
   hostname: {
     get () {
       return this.headers.host || this.headers[':authority']