--- a/lib/route.js
+++ b/lib/route.js
@@ -1,7 +1,6 @@
 'use strict'
 
 const FindMyWay = require('find-my-way')
-const proxyAddr = require('proxy-addr')
 const Context = require('./context')
 const handleRequest = require('./handleRequest')
 const { hookRunner, hookIterator, lifecycleHooks } = require('./hooks')
@@ -48,7 +47,6 @@ function buildRouting (options) {
 
   let avvio
   let fourOhFour
-  let trustProxy
   let requestIdHeader
   let querystringParser
   let requestIdLogLabel
@@ -56,7 +54,6 @@ function buildRouting (options) {
   let hasLogger
   let setupResponseListeners
   let throwIfAlreadyStarted
-  let proxyFn
   let genReqId
   let disableRequestLogging
   let ignoreTrailingSlash
@@ -74,8 +71,6 @@ function buildRouting (options) {
       setupResponseListeners = fastifyArgs.setupResponseListeners
       throwIfAlreadyStarted = fastifyArgs.throwIfAlreadyStarted
 
-      proxyFn = getTrustProxyFn(options)
-      trustProxy = options.trustProxy
       requestIdHeader = options.requestIdHeader
       querystringParser = options.querystringParser
       requestIdLogLabel = options.requestIdLogLabel
@@ -302,17 +297,6 @@ function buildRouting (options) {
     }
 
     var id = req.headers[requestIdHeader] || genReqId(req)
-    var hostname = req.headers.host || req.headers[':authority']
-    var ip = req.connection.remoteAddress
-    var ips
-
-    if (trustProxy) {
-      ip = proxyAddr(req, proxyFn)
-      ips = proxyAddr.all(req, proxyFn)
-      if (ip !== undefined && req.headers['x-forwarded-host']) {
-        hostname = req.headers['x-forwarded-host']
-      }
-    }
 
     var loggerOpts = {
       [requestIdLogLabel]: id,
@@ -327,7 +311,7 @@ function buildRouting (options) {
 
     var queryPrefix = req.url.indexOf('?')
     var query = querystringParser(queryPrefix > -1 ? req.url.slice(queryPrefix + 1) : '')
-    var request = new context.Request(id, params, req, query, req.headers, childLogger, ip, ips, hostname)
+    var request = new context.Request(id, params, req, query, req.headers, childLogger)
     var reply = new context.Reply(res, context, request, childLogger)
 
     if (disableRequestLogging === false) {
@@ -419,27 +403,6 @@ function preParsingHookRunner (functions, request, reply, cb) {
   }
 
   next(null, request[kRequestPayloadStream])
-}
-
-function getTrustProxyFn (options) {
-  const tp = options.trustProxy
-  if (typeof tp === 'function') {
-    return tp
-  }
-  if (tp === true) {
-    // Support plain true/false
-    return function () { return true }
-  }
-  if (typeof tp === 'number') {
-    // Support trusting hop count
-    return function (a, i) { return i < tp }
-  }
-  if (typeof tp === 'string') {
-    // Support comma-separated tps
-    const vals = tp.split(',').map(it => it.trim())
-    return proxyAddr.compile(vals)
-  }
-  return proxyAddr.compile(tp || [])
 }
 
 module.exports = { buildRouting, validateBodyLimitOption }