In this PR I'm moving `ip`, `ips` and `hostname` to be getters instead of fields on the request as suggested my @mcollina in https://github.com/fastify/fastify/issues/2416

As part of that I extracted the tests on the request object from `handleRequest.test.js`, moved them to their own file and expanded them.

I've run the benchmarks and I can't see any significant change.

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [ ] commit message and code follows [Code of conduct](https://github.com/fastify/fastify/blob/master/CODE_OF_CONDUCT.md)